# swagger-api.json
{"swagger":"2.0","info":{"version":"1.0.0","title":"Bill Manager API"},"host":"localhost:8095","basePath":"/","tags":[{"name":"bill-manager-resource","description":"Bill Manager Resource"}],"paths":{"/api/bill/add-expense":{"post":{"tags":["bill-manager-resource"],"summary":"API to add expenses for friends.","operationId":"addExpenseUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"expenseRequest","description":"expenseRequest","required":true,"schema":{"$ref":"#/definitions/ExpenseRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ExpenseResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/bill/delete-expense":{"delete":{"tags":["bill-manager-resource"],"summary":"API to delete expenses for friends.","operationId":"deleteExpenseUsingDELETE","produces":["*/*"],"parameters":[{"in":"body","name":"expenseRequest","description":"expenseRequest","required":true,"schema":{"$ref":"#/definitions/ExpenseRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ExpenseResponse"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/api/bill/edit-expense":{"put":{"tags":["bill-manager-resource"],"summary":"API to edit expenses for friends.","operationId":"editExpenseUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"expenseRequest","description":"expenseRequest","required":true,"schema":{"$ref":"#/definitions/ExpenseRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ExpenseResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/bill/get-expenses":{"get":{"tags":["bill-manager-resource"],"summary":"API to pull the expenses for all the friends.","operationId":"getExpenseUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ExpenseResponse"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/bill/split-expense":{"post":{"tags":["bill-manager-resource"],"summary":"API to split the expenses between friends and display the settlement details.","operationId":"splitExpenseUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"expenseRequest","description":"expenseRequest","required":true,"schema":{"$ref":"#/definitions/ExpenseRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/SplitExpenseResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"Bill":{"type":"object","properties":{"exepenseAmount":{"type":"number","description":"Expense Amount"},"expenseDescription":{"type":"string","description":"Expense Description"},"expenseID":{"type":"integer","format":"int32","description":"Expense ID"}},"title":"Bill"},"ExpenseRequest":{"type":"object","properties":{"friends":{"type":"array","description":"Friends sharing the expenses","items":{"$ref":"#/definitions/Friend"}}},"title":"ExpenseRequest"},"ExpenseResponse":{"type":"object","properties":{"friends":{"type":"array","description":"Friends List Response","items":{"$ref":"#/definitions/Friend"}}},"title":"ExpenseResponse"},"Friend":{"type":"object","properties":{"bill":{"$ref":"#/definitions/Bill"},"emailAddress":{"type":"string","description":"Friend's email address"},"expenseID":{"type":"integer","format":"int32"},"name":{"type":"string","description":"Friend's full name"}},"title":"Friend"},"SplitExpenseResponse":{"type":"object","properties":{"expenseResponseModelList":{"type":"array","description":"Expense List Response","items":{"$ref":"#/definitions/SplitExpenseResponseModel"}}},"title":"SplitExpenseResponse"},"SplitExpenseResponseModel":{"type":"object","properties":{"emailAddress":{"type":"string","description":"Friend's email Address"},"friendName":{"type":"string","description":"Friend's full Name"},"setledExpenses":{"type":"number","description":"settled Expense Value"},"settlementDescription":{"type":"string","description":"settlement Description"}},"title":"SplitExpenseResponseModel"}}}